Redesigned the scripts, its now the Task Runner and Builder systems.
