+++ packages/portfolio/blog/designing-a-web-app/package.json
@@ -14,7 +14,7 @@
   ],
   "foil": {
     "title": "Designing a Web Application",
-    "main": "blog/main.js",
+    "main": "../main.tsx",
     "datePublished": "2016-09-01T00:30:00.000Z"
   }
 }
