+++ packages/portfolio/blog/intro-sound-programming/package.json
@@ -11,7 +11,7 @@
   ],
   "foil": {
     "title": "Intro to Sound Programming",
-    "main": "blog/main.js",
+    "main": "../main.tsx",
     "datePublished": "2017-02-05T00:30:00.000Z"
   }
 }
