+++ packages/portfolio/blog/javascript-at-2017/package.json
@@ -14,7 +14,7 @@
   ],
   "foil": {
     "title": "JavaScript @ 2017",
-    "main": "blog/main.js",
+    "main": "../main.tsx",
     "datePublished": "2017-03-11T01:54:43.049Z"
   }
 }
