+++ packages/portfolio/scripts/lib.ts
@@ -1,5 +1,5 @@
 /**
- * 🔳 Alain.xyz Builder
+ * 🔳 Alain.xyz Task Runner
  * Compiles the portfolio and populates it in MongoDB. 
  * Each portfolio item has a unique Build step.
  */
@@ -11,22 +11,27 @@ async function main() {
 
   // Start Build Process
   console.log(
-    cyan('🔳 Alain.xyz Builder v0.5.0')
+    cyan('🔳 Alain.xyz Task Runner v0.6.0')
   )
 
   // Run each task
   var scripts = Object.values(buildScripts);
 
   for (var i = 0; i < scripts.length; i++) {
-    console.log(`\n👟 ${gray(`Running Task (${i + 1}/${scripts.length})...`)}\n`);
+
+    let progress = `(${i + 1}/${scripts.length})`;
+
+    console.log(`\n👟 ${gray(`Running Task ${progress}...`)}\n`);
+
     await scripts[i]()
-    .then( res => {
-      console.log(`\n✔️️ ${green(`Finished Task (${i + 1}/${scripts.length})!`)}\n`);
-    })
-    .catch(err => {
-      console.log(`\n❌ ${red(`Failed Task (${i + 1}/${scripts.length})!`)}\n`);
-      console.error(err);
-    });
+      .then(res => {
+        console.log(`\n✔️️ ${green(`Finished Task ${progress}!`)}\n`);
+      })
+      .catch(err => {
+        console.log(`\n❌ ${red(`Failed Task ${progress}!`)}\n`);
+        console.error(err);
+      });
+
   }
   console.log('\n💮 ' + gray(`Finished processing ${scripts.length} tasks!\n`));
