+++ packages/portfolio/scripts/tasks/builder/loaders/index.ts
@@ -1,0 +1,12 @@
+export * from './ts';
+export * from './md';
+export * from './mp3';
+
+export type Loader = {
+    test: Object,
+    loader: () => any;
+}
+
+export type Foil = {
+    
+}
