+++ packages/portfolio/scripts/tasks/builder/loaders/md.ts
@@ -1,0 +1,14 @@
+import markademic from 'markademic';
+import { join } from 'path';
+import { readFileSync } from 'fs';
+
+export let md = {
+    test: { file: /\.md$/ },
+    loader: foil => ({
+      ...foil,
+      data: markademic({
+        input: readFileSync(foil.file).toString(),
+        rerouteLinks: (link) => join(foil.permalink, link)
+      })
+    })
+  }
