+++ packages/portfolio/scripts/tasks/index.ts
@@ -1,3 +1,2 @@
 export * from './clean';
-export * from './subapp';
-export * from './package';
+export * from './builder';
